feat: add HelpServicesPage component with service grid and request form
