+++ src/App.tsx
@@ -12,7 +12,7 @@ import AdditionalServices from "@/components/AdditionalServicesPage/index";
 import AccountPage from "@/components/AccountPage/index";
 import SignUpPage from "@/components/SignUp-page";
 import Confirmation from "@/components/Confirmation/index";
-import HelpServicesPage from "@/components/Help-Services-page";
+import HelpServicesPage from "@/components/HelpPage/index";
 import FoodPage from "@/components/FoodPage/index";
 import ActiveReservationPage from '@/components/ActiveReservationPage/index';
