+++ src/i18n/locales/en/translation.json
@@ -99,6 +99,7 @@
     "reservations": "Reservations"
   },
   "help": {
+    "title": "Help & Services",
     "help": "Help & Services",
     "food": "Drinks, Food & Snacks",
     "wifiSupport": "Wifi Support",
