+++ src/i18n/locales/es/translation.json
@@ -117,6 +117,7 @@
     "reservations": "Reservas"
   },
   "help": {
+    "title": "Ayuda y Servicios",
     "help": "Atención al cliente",
     "food": "Bebidas, Comida y Snacks",
     "wifiSupport": "Soporte de Wifi",
